requirements.txt
google-ads==27.0.0
requests
